Create Final Project
Truth or Dare is a game I loved to play with my friends in grade school. I found it to be exciting that I can apply what I have learned in this class to a game that everyone knows. I have never been the best at computer work like python, but I found this software to be fun and interesting. I still play "Truth or Dare" with my fiancé on long car rides to keep our minds busy and entertained as well as with my sisters five children. Whether its "you have to answer this truthfully" or "Go do this specific thing", its an interesting and possibly funny game. After putting this code together the well known truth or dare is not a decision the player gets to make. Its in the hands of the python game. This class is like no other I have taken before. Python is a cool way to create and enjoy!